jupiter-controller-fw: Add script to generate update data
In the next commit, the script will be changed to refer to the data
generated by this script.

The goal here is to make this part of the update process in some for,
so that there is no way to forget to update the values or use the wrong
values.

Signed-off-by: Samuel Dionne-Riel <samuel@dionne-riel.com>